+++ README.md
@@ -24,22 +24,3 @@ Alternatively, you can run Jupyter Server and visiting the `/tree` endpoint:
 ```
 > jupyter server
 ```
-
-## Further Details
-
-This project also includes an API for shimming traits that moved from `NotebookApp` in to `ServerApp` in Jupyter Server. This can be used by applications that subclassed `NotebookApp` to leverage the Python server backend of Jupyter Notebooks. Such extensions should *now* switch to `ExtensionApp` API in Jupyter Server and add `NBClassicConfigShimMixin` in their inheritance list to properly handle moved traits.
-
-For example, an application class that previously looked like:
-```python
-from notebook.notebookapp import NotebookApp
-
-class MyApplication(NotebookApp):
-```
-should switch to look something like:
-```python
-from jupyter_server.extension.application import ExtensionApp
-from nbclassic.shim import NBClassicConfigShimMixin
-
-class MyApplication(NBClassicConfigShimMixin, ExtensionApp):
-```
-
