--- nbclassic/tests/confs/jupyter_my_ext_config.py
@@ -1,1 +1,0 @@
-c.MyExt.hello = 'My extension'
