--- nbclassic/tests/confs/jupyter_notebook_config.py
@@ -1,3 +1,0 @@
-c.NotebookApp.allow_credentials = False
-c.NotebookApp.port = 8889
-c.NotebookApp.password_required = True
