--- nbclassic/tests/confs/jupyter_server_config.py
@@ -1,1 +1,0 @@
-c.ServerApp.port = 7774
