+++ setup.cfg
@@ -28,6 +28,7 @@ packages = find:
 python_requires = >=3.6
 install_requires =
     notebook<7
+    notebook_shim>=0.1.0
     jupyter_server~=1.8
 
 [options.data_files]
